+++ PRIME@UCR.Application/Implementations/UserAdministration/PerteneceService.cs
@@ -1,4 +1,7 @@
-﻿using PRIME_UCR.Application.Repositories.UserAdministration;
+﻿using Microsoft.AspNetCore.Authorization;
+using Microsoft.AspNetCore.Components.Authorization;
+using PRIME_UCR.Application.DTOs.UserAdministration;
+using PRIME_UCR.Application.Repositories.UserAdministration;
 using PRIME_UCR.Application.Services.UserAdministration;
 using System;
 using System.Collections.Generic;
@@ -11,19 +14,35 @@ namespace PRIME_UCR.Application.Implementations.UserAdministration
     {
         private readonly IPerteneceRepository _perteneceRepository;
 
-        public PerteneceService(IPerteneceRepository perteneceRepository)
+        private readonly IAuthorizationService authorizationService;
+
+        private readonly AuthenticationStateProvider authenticationStateProvider;
+
+        public PerteneceService(IPerteneceRepository perteneceRepository,
+            IAuthorizationService _authorizationService,
+            AuthenticationStateProvider _authenticationStateProvider)
         {
             _perteneceRepository = perteneceRepository;
+            authorizationService = _authorizationService;
+            authenticationStateProvider = _authenticationStateProvider;
         }
 
         public async Task DeleteUserOfProfileAsync(string idUser, string idProfile)
         {
-            await _perteneceRepository.DeleteUserFromProfileAsync(idUser, idProfile);
+            var user = (await authenticationStateProvider.GetAuthenticationStateAsync()).User;
+            if( (await authorizationService.AuthorizeAsync(user, AuthorizationPolicies.CanManageUsers.ToString())).Succeeded)
+            {
+                await _perteneceRepository.DeleteUserFromProfileAsync(idUser, idProfile);
+            }
         }
 
         public async Task InsertUserOfProfileAsync(string idUser, string idProfile)
         {
-            await _perteneceRepository.InsertUserToProfileAsync(idUser, idProfile);
+            var user = (await authenticationStateProvider.GetAuthenticationStateAsync()).User;
+            if ((await authorizationService.AuthorizeAsync(user, AuthorizationPolicies.CanManageUsers.ToString())).Succeeded)
+            {
+                await _perteneceRepository.InsertUserToProfileAsync(idUser, idProfile);
+            }
         }
     }
 }
